+++ lib/oakdex/pokedex/item.rb
@@ -6,6 +6,11 @@ module Oakdex
     class Item < Base
       json_folder 'item'
       translate :names, :name
+
+      def description(locale = 'en')
+        translations = descriptions.last['translations']
+        translations[locale] || translations['en']
+      end
     end
   end
 end